pyproject.toml
[project]
name = "kafka-flink-tool"
version = "0.1.0"
description = "Kafka to Flink to Hologres automation tool"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0,<7.0",
    "kafka-python>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "click>=8.0.0,<9.0.0",
]

[project.scripts]
kafka-flink-tool = "kafka_flink_tool.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
